SimpleZoneCharts
Simple zone charts based on three zones on the football field: redzone (1 to 20 yards), midfield (20 to -20 yards), and backed up (-20 to 1 yards). For each zone, displays the amounts and percentages of runs vs plays, most and least used formations, most and least efficient formations, and quarters most in the zone.